noun-project-icons/requirements.txt
python-thenounproject
